feat: rename file-utils to file
