+++ api/routes/convert-routes.js
@@ -5,26 +5,15 @@ import {
   UPLOAD_DIR,
   REDIS_CONNECTION_CONFIG,
 } from "./../config/config.js"
-import {
-  checkFoldersExistAsync,
-  getAllFilesInDir,
-} from "./../utils/file-utils.js"
+import { checkFoldersExistAsync, getAllFilesInDir } from "../utils/file.js"
 import logger from "./../utils/logger.js"
 
 const flowProducer = new FlowProducer({ connection: REDIS_CONNECTION_CONFIG })
 
 const router = express.Router()
 
-// TODO: convert to post and only allow if correct password is provided that the user got via mail OR add the password automatically in the link
 router.post("/convert/files", async (req, res, next) => {
   const { jobId, subject, email, city, teacher } = req.body
-  console.log("/convert/files req.body", {
-    jobId,
-    subject,
-    email,
-    city,
-    teacher,
-  })
 
   if (!jobId || !subject || !email || !city || !teacher) {
     const error = new Error("Missing required fields")
